bugfix CWE-120
set limit to strncat
