+++ example/Makefile.dpdk.in
@@ -17,7 +17,7 @@ include $(RTE_SDK)/mk/rte.vars.mk
 APP = ndpiReader.dpdk
 LIBNDPI = $(PWD)/../src/lib/libndpi.a
 
-SRCS-y := reader_util.c intrusion_detection.c dns_utils.c ndpiReader.c
+SRCS-y := reader_util.c intrusion_detection.c ndpiReader.c
 
 CFLAGS += -g
 CFLAGS += -Wno-strict-prototypes -Wno-missing-prototypes -Wno-missing-declarations -Wno-unused-parameter -I $(PWD)/../src/include @CFLAGS@ -DUSE_DPDK
