+++ example/Makefile.in
@@ -7,7 +7,7 @@ LIBNDPI=$(SRCHOME)/lib/libndpi.a
 LDFLAGS=$(LIBNDPI) @PCAP_LIB@ @LIBS@ @ADDITIONAL_LIBS@ -lpthread -lm @LDFLAGS@
 HEADERS=intrusion_detection.h reader_util.h dns_utils.h $(SRCHOME)/include/ndpi_api.h \
         $(SRCHOME)/include/ndpi_typedefs.h $(SRCHOME)/include/ndpi_protocol_ids.h
-OBJS=ndpiReader.o reader_util.o dns_utils.o intrusion_detection.o
+OBJS=ndpiReader.o reader_util.o intrusion_detection.o
 PREFIX?=@prefix@
 
 ifneq ($(BUILD_MINGW),)
@@ -27,7 +27,7 @@ libndpiReader.a: $(COMMON_SOURCES:%.c=%.o) $(LIBNDPI)
 ndpiReader: libndpiReader.a $(LIBNDPI) ndpiReader.o
 	$(CC) $(CFLAGS) ndpiReader.o libndpiReader.a -o $@ $(LDFLAGS)
 
-ndpiSimpleIntegration: ndpiSimpleIntegration.o dns_utils.o $(LIBNDPI) 
+ndpiSimpleIntegration: ndpiSimpleIntegration.o $(LIBNDPI) 
 	$(CC) $(CFLAGS) $? -o $@ $(LDFLAGS)
 
 %.o: %.c $(HEADERS) Makefile
